requirements.txt
streamlit
tensorflow
numpy
pillow
streamlit-drawable-canvas
scikit-learn